Dodanie zabezpieczenia dla TypuTagu (przypisany do tagu nie mozna usunac), grupy (przypisana do uzytkownika nie mozna usunac).

+++ Portal/src/main/java/com/portal/controller/TagTypeController.java
@@ -1,25 +1,22 @@
 package com.portal.controller;
 
+import com.portal.dao.interfaces.TagDAOI;
 import com.portal.dao.interfaces.TagTypeDAOI;
 import com.portal.entity.TagType;
-import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletResponse;
+import java.util.List;
 
 @Controller
 public class TagTypeController {
     @Autowired
     public TagTypeDAOI tagTypeDAO;
+    @Autowired
+    public TagDAOI tagDAO;
     
     @RequestMapping(value="/type", method= RequestMethod.GET)
     @ResponseBody
@@ -47,7 +44,11 @@ public class TagTypeController {
 
     @RequestMapping(value="type/{id}", method= RequestMethod.DELETE)
     public void delete(@PathVariable Long id, HttpServletResponse response){
-        tagTypeDAO.delete(id);
-        response.setStatus(HttpServletResponse.SC_OK);
+        if (tagDAO.existsForType(tagTypeDAO.get(id))) {
+            response.setStatus(HttpServletResponse.SC_CONFLICT);
+        } else {
+            tagTypeDAO.delete(id);
+            response.setStatus(HttpServletResponse.SC_OK);
+        }
     }
 }
